Add ability to opt in/out of pickup events
